+++ app/handlers.go
@@ -103,7 +103,7 @@ func addHandler(w http.ResponseWriter, r *http.Request) {
 	}
 
 	s := script{
-		Id:    time.Now().Unix(),
+		Id:    float64(time.Now().Unix()),
 		Meta:  "",
 		Title: title,
 		Url:   uri,
